Adds some templates for the suggested harvesting workflow
Adds some templates to work through the suggested harvester workflow.

Step 1 is adding a title and a URL, which then submits to a page telling
the user we have identified the catalogue and showing a list of
organisations from which they can harvest - note there is also a page at
/harvest/unknown which asks the user to specify which catalogue they
use.

The user is then asked how frequently they wish to have the harvester
run, then whether they want notifications after each run.

And that's it.