Add a Question X of Y section next to back button in flows.
For each flow (of more than one page) we add a Question x of y section
next to the back button to show our current position in the flow.

Would be nice to be able to abstract out the total number of steps
(perhaps to a local layout?) so that we don't have to update it on each
page when we add a new one - given that inserting a new page will
probably change all of the x values, probably no rush.