Mocks up a WMS preview to be used in-line with the dataset
Rather than have an entirely different place to go to preview layers on
a map (or as well as), and because we've had feedback that the boundary
view on the dataset page isn't useful, this shows what an inline preview
of WMS might look like.

Can be dropped into relevant page with

{% include "includes/wms_preview.html" %}

and may be of use in the dataset preview page as we consider the display
of UKLP datasets.